Wikitext link support (#6)
* add a stub for a url report

* Provide a simple boring report of outgoing URLs

* Implemented a forward link report

* Support wikilinks

Updated goldmark version to make sure that the wikilink extension
works. Added that as a dependency. Use it to extract and log the
wikilink-style links.

* Refactoring for wikitext improvements

This CL restructures the code to better support wikitext style
links. The reporting facilities for wikitext style links are
temporarily unavailable.

* Improved wikitext link handling

* Restore wikilink reporting

Now with working and correctly formatted wikilink reporting.